# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoundApp OpenApi Documentation",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "francesco.simeoli@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://roundapp.local/api",
            "description": "RoundApp API Server"
        },
        {
            "url": "https://roundapp.local"
        }
    ],
    "paths": {
        "/auth/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Password reset request",
                "description": "Password reset request",
                "operationId": "authPasswordForgot",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@roundapp.local"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "We have emailed your password reset link!"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (invalid email | email not registered)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "json"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "The given data was invalid.",
                                        "errors": {
                                            "email": "We can't find a user with that email address."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request OAuth2 token flow password",
                "description": "Request OAuth2 token flow password",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@roundapp.local",
                                    "password": "roundapptest"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "type": "string"
                                        },
                                        "verified": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "token_type": "Bearer",
                                        "expires_id": 31535999,
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNDkyNDlmMTkzOTJlY2RmZWFjOWZlN2NhMjY5NDYxNjhmNzdmZjExOGYxZjJiNmNkZGFjNTRkNjM1ODAxMmRlNWVhNGIxOTMyMDBmZDNkZTAiLCJpYXQiOjE1OTI1ODA0MDcsIm5iZiI6MTU5MjU4MDQwNywiZXhwIjoxNjI0MTE2NDA2LCJzdWIiOiIxMSIsInNjb3BlcyI6WyIqIl19.DL4Wx6oteW00Af-U5uSzdNjACNkZoYdypiXATP1YsSWA77smZnOMlRuQbg4H2TYzLH-Vw5bPB7t4Q2PBOmG7Afkm6RuKgTZMLXE2HNcaYlQRdtY_Mywcm1wJp-g6A5re-0kEQwWvhyZzlC4hnNdww-lG2ofclgn5dLV6VGZ65ARUedNYRbVNX0LAyNTUgeke4BKfpyeHMNdmAcgrpwdDgBelbvBz_fl05EmF3dFhOR55nrY6Obh2BFFX5SmiWztCIsY4blOR04JitSiLsf1aBRHE_xoAdmBAR-Kq9jDfCfx7-l3371SlUgzmwRQ1NtGfjNMgCdJe1XjNidqob-VR2hW9s29rozcnt4UUVGCHRopltYpQv7N8500PU4og7pUJ9V-VNxKxjDww3y0Bl-54yg_DRwpdrCcT-wwGdtdDn5FF041k_LrV8j8Wbap0JdeWKOCc96zZXJUpELcikFq-NAN66D5dpCt6r0RZsZAhvNaqanleDTaJzbcHZ9z",
                                        "refresh_token": "def502006860d206e0b8f88e208d3fd9dbfd55dacad90c03f6276334642aa6213b3e3a541889583d90e301c566bc2d4e0baf5e8733e59a284bc177e2951ea2733775f64b8ecebd2277edac4fbf8e6a4872d6940a9ada4c1ff7ffece5631f3742f9a1c9368ec9d3dddf85c303e87d4143226cf094e4bca99bfb39b987e921b8674071bebd1cd23a17e8a5545ef767bcac3cb7fc3bb9fcb19c80ec8a7d19d407d42084e31d56a47a95e8c80244b7ac0e3ba7c95c35bf80230d11df45690543ca3cef315821c1c8f8fa11e4f32cf8cf6ad06b6ca9211f32873dd902c369a900ddb482cc780b1522565ddaac0ca05a7039011605276b1d64a0e34067e987e38849d109bbd660fedf12e76a03f359f3cf9e2b99b03a535209d8b60ae9ae87697407021667605b4ce0b185ed6fd7888c2d0b95e91210a8f4f16dfa2e1135ee7d3f88fc5af969ad9b3099e4e9794d9b81a60a2720582b61a17157627cfca1ebd8f7c6815ffa465dd9",
                                        "verified": false
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "error_description": {
                                            "type": "string"
                                        },
                                        "hint": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "error": "invalid_grant",
                                        "error_description": "The provided authorization grant (e.g., authorization code, resource owner credentials) or refresh token is invalid, expired, revoked, does not match the redirection URI used in the authorization request, or was issued to another client.",
                                        "hint": "",
                                        "message": "The provided authorization grant (e.g., authorization code, resource owner credentials) or refresh token is invalid, expired, revoked, does not match the redirection URI used in the authorization request, or was issued to another client."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://roundapp.local/oauth/authorize",
                        "tokenUrl": "https://roundapp.local/oauth/token",
                        "refreshUrl": "https://roundapp.local/token/refresh",
                        "scopes": [],
                        "client_id": "1",
                        "secret": "yo4kJNkiGGIiU0YqZ5jP1ucCArLM11FC7z5QdEFY"
                    }
                }
            }
        }
    },
...
}